Move status code to model.py and rename types
Per the review discussion we decided that:

* status-related types will be placed in op/model.py;
* those types will have a "Status" postfix, e.g.: ActiveStatus.
